ref(components): Move non-Shadcn UI components
- Moved all non-Shadcn UI components to "components/custom/ui" directory
  - This also make sure the ESLint includes that directory but not the
    "components/ui"
